+++ python/par_term_emu_core_rust/__init__.py
@@ -73,7 +73,7 @@ except ImportError:
     encode_client_message = None
     decode_client_message = None
 
-__version__ = "0.18.1"
+__version__ = "0.23.0"
 __all__ = [
     "Attributes",
     "CursorStyle",